Added Clock Syncronization
Added Clock Syncronization to fix interpolation issues on devices with a different OS time.

Refactored the lobby to be less hard-coded and made the text box set itself as your global IP.

Added Boolean functionality to the networksyncer along with setting the updates per second and percent of updates that are required. This will reduce bandwidth issues with many devices.

Added a function to create players.

Various bug fixes.