Nested Syncer Update
Network Syncers now support being nested inside a single object. For my project, this can be used to have a player object sync it's rotation and translation, and have the weapon manager have it's own syncer for readability.

This also allows for more data to be sent over different nodes.